move comptoken bankrun tests into subdirectory
