+++ test/compto-test-client/comptoken-tests/dailyDistributionEvent.js
@@ -2,9 +2,9 @@ import { SYSVAR_SLOT_HASHES_PUBKEY, Transaction, TransactionInstruction } from "
 import { Clock, start } from "solana-bankrun";
 
 import { TOKEN_2022_PROGRAM_ID } from "@solana/spl-token";
-import { get_default_comptoken_mint, get_default_global_data, get_default_unpaid_interest_bank, get_default_unpaid_ubi_bank, GlobalDataAccount, MintAccount, programId, TokenAccount } from "./accounts.js";
-import { Assert } from "./assert.js";
-import { DEFAULT_ANNOUNCE_TIME, DEFAULT_DISTRIBUTION_TIME, DEFAULT_START_TIME, Instruction, SEC_PER_DAY } from "./common.js";
+import { get_default_comptoken_mint, get_default_global_data, get_default_unpaid_interest_bank, get_default_unpaid_ubi_bank, GlobalDataAccount, MintAccount, TokenAccount } from "../accounts.js";
+import { Assert } from "../assert.js";
+import { compto_program_id_pubkey, DEFAULT_ANNOUNCE_TIME, DEFAULT_DISTRIBUTION_TIME, DEFAULT_START_TIME, Instruction, SEC_PER_DAY } from "../common.js";
 
 async function test_dailyDistributionEvent() {
     let comptoken_mint = get_default_comptoken_mint();
@@ -13,7 +13,7 @@ async function test_dailyDistributionEvent() {
     let interest_bank = get_default_unpaid_interest_bank();
     let ubi_bank = get_default_unpaid_ubi_bank();
     const context = await start(
-        [{ name: "comptoken", programId }],
+        [{ name: "comptoken", programId: compto_program_id_pubkey }],
         [
             comptoken_mint.toAccount(),
             global_data.toAccount(),
@@ -42,7 +42,7 @@ async function test_dailyDistributionEvent() {
 
     let data = Buffer.from([Instruction.DAILY_DISTRIBUTION_EVENT])
 
-    const ixs = [new TransactionInstruction({ programId, keys, data })];
+    const ixs = [new TransactionInstruction({ programId: compto_program_id_pubkey, keys, data })];
     const tx = new Transaction();
     tx.recentBlockhash = blockhash;
     tx.add(...ixs);
